Refactor : addCardContainer를 addCardPage로 변경
