+++ index.js
@@ -3,8 +3,6 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 import App from "./src/App";
 
-import AddCardContainer from "./src/components/AddCard/AddCardContainer";
-
 reactDom.render(
   <BrowserRouter>
     <App />
