--- src/components/AddCard/AddCardContainer/index.js
@@ -1,1 +1,0 @@
-export { default } from "./AddCardContainer";
